Refactor: type 관련 파일 src/types로 이동
